+++ drop-down-terminal@gs-extensions.zzrough.org/terminal.js
@@ -276,7 +276,11 @@ const DropDownTerminal = new Lang.Class({
         let terminal = new Vte.Terminal();
 
         terminal.set_can_focus(true);
-        terminal.set_background_transparent(false);
+
+        if (terminal.set_background_transparent) { // removed in 0.34.8
+                terminal.set_background_transparent(false);
+        }
+
         terminal.set_allow_bold(true);
         terminal.set_scroll_on_output(true);
         terminal.set_scroll_on_keystroke(true);
@@ -392,11 +396,18 @@ const DropDownTerminal = new Lang.Class({
         let transparencyLevel = this._settings.get_uint(TRANSPARENCY_LEVEL_SETTING_KEY) / 100.0;
         let hasScrollbar = this._settings.get_boolean(SCROLLBAR_VISIBLE_SETTING_KEY);
 
-        this._terminal.set_background_image(null); // required to update the opacity after realize
+        if (this._terminal.set_background_image) { // remove in 0.34.8
+            this._terminal.set_background_image(null); // required to update the opacity after realize
+        }
 
         if (this._terminalScrollbar.set_opacity) { // 3.7.10+
             // starting from 3.7.10, all gtk widgets can have their opacity changed
-            this._terminal.set_opacity((isTransparent ? transparencyLevel : 1.0) * 0xffff);
+            if (this._terminal.set_background_image) { // remove in 0.34.8
+                this._terminal.set_opacity((isTransparent ? transparencyLevel : 1.0) * 0xffff);
+            } else {
+                this._terminal.opacity = transparencyLevel;
+            }
+
             this._terminalScrollbar.set_opacity(isTransparent ? transparencyLevel : 1.0);
         } else {
             // making the window transparent also makes the font transparent which make it a bit more difficult