+++ NEWS.md
@@ -59,4 +59,8 @@ v9 (2013-03-25)
 v10 (2013-04-21)
 - bugfix: [Allow a block or underline cursor in the terminal (github #36)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/36)
 - bugfix: [Switch focus faster (github #38)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/38)
-- code: get rid of the gtk allocation regression fix for gtk >= 3.8.1
+
+v11 (2013-04-28)
+- bugfix: [Height change doesn''t work on 3.8.1 (github #41)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/41)
+- bugfix: [Error upgrading extension (github #40)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/40)
+- code: make the gtk allocation regression fix conditional ; the story is not finished yet see b.g.o. #696882