+++ NEWS.md
@@ -64,3 +64,6 @@ v11 (2013-04-28)
 - bugfix: [Height change doesn''t work on 3.8.1 (github #41)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/41)
 - bugfix: [Error upgrading extension (github #40)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/40)
 - code: make the gtk allocation regression fix conditional ; the story is not finished yet see b.g.o. #696882
+
+v12 (2013-05-08)
+- bugfix: [Height change doesn''t work on 3.8.1 (github #41)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/41)