+++ NEWS.md
@@ -67,3 +67,5 @@ v11 (2013-04-28)
 
 v12 (2013-05-08)
 - bugfix: [Height change doesn''t work on 3.8.1 (github #41)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/41)
+
+v13 (????-??-??)